(atlassian) add quickInsert & typeAhead plugins that almost work
This includes a massive amount of plumbing and scripts as copied from the Atlassian git repo.
Also a few external dependencies to @atlaskit/icon & item which I saw too burdensome to unravel
to this repo. Especially them Item library is annoying, as it has been written in JavaScript,
for who knows what reason, and which doesnt even include types.

The main contributions of this commit are the aforementioned plugins which include adding the
ProviderFactory, contentComponents and WithPluginState HOC. Currently the editor crashes when
the typeAheadPlugin is activated via / which displays the quickInsert menu and a mouse is moved.
The next commit should aim to fix it.